improve plugin routing and update user module
